Alt-ergo

The Alt-ergo prover

The alt-ergo package installs the alt-ergo binary, whose documentation is available through the --help option.

Command Line Parsing

The command line parsing is computed in our common internal library (see Alt_ergo_common) using cmdliner

Plugins

Why3 parser

for the Why3 parser plugin documentation see ABWhy3

Fourier Motzkin simplex

for the Fourier Motzkin simplex plugin documentation see FMsimplex