Altgr-ergo

The Altgr-ergo prover

The altgr-ergo package installs the altgr-ergo binary, which is a graphical interface to the alt-ergo automated theorem prover.

Command Line Parsing

The command line parsing is computed in our common internal library (see Alt_ergo_common) using cmdliner