Alt-ergo

The Alt-ergo prover

The alt-ergo package installs the alt-ergo binary, whose documentation is available through the --help option.

Alt_Ergo_common

This package uses the Alt-Ergo_common internal lib (see Alt_ergo_common) for parsing command line, input file, and main solving loop.

Plugins

Why3 parser

for the Why3 parser plugin documentation see ABWhy3

Fourier Motzkin simplex

for the Fourier Motzkin simplex plugin documentation see FMsimplex